Refactor: Rename UserDto to ResponseUserDto in UserService
Changed the name of UserDto to ResponseUserDto in the UserService.java class. This name change more accurately reflects the object's role as a response object. This helps to increase code readability and comprehension by making clear distinctions between data transfer objects based on their purpose and usage in the application.