Rename and refactor UserJpaRepositoryTest to UserRepositoryTest
The file UserJpaRepositoryTest has been renamed to UserRepositoryTest. In addition, the test has been updated to use Mockito for mocking dependencies, and now includes a new test case for the method "getUsers". The import statement has also been updated, replacing the use of TestUtil with TestUtils for better naming consistency.