Update UserMapper for User record usage
The UserMapper class was updated to correspond with changes in the User class. The User class was transformed from a normal Java class to a record which resulted in changes to how its attributes are accessed. The changes in the UserMapper class now reflect the usage of methods to access these attributes rather than directly calling them.